feat: User specific content with JWT claims

+++ Controllers/v1/PostsController.cs
@@ -1,10 +1,11 @@
 using blog_api_jwt.Contracts.v1;
-using blog_api_jwt.Domain;
-using Microsoft.AspNetCore.Mvc;
-using System;
 using blog_api_jwt.Contracts.v1.Requests;
 using blog_api_jwt.Contracts.v1.Responses;
+using blog_api_jwt.Domain;
 using blog_api_jwt.Services.Interfaces;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace blog_api_jwt.Controllers.v1;
@@ -12,6 +13,7 @@ namespace blog_api_jwt.Controllers.v1;
 public class PostsController : Controller
 {
     private readonly IPostService _postService;
+    private readonly UserManager<User> _userManager;
 
     public PostsController(IPostService postService)
     {
@@ -21,9 +23,12 @@ public class PostsController : Controller
     [HttpPost(ApiRoutes.Posts.Create)]
     public async Task<IActionResult> Create([FromBody] CreatePostRequest request)
     {
+        var userId = HttpContext.User.Claims.SingleOrDefault(u => u.Type == "id")?.Value;
+
         var post = new Post
         {
-            Name = request.Name
+            Name = request.Name,
+            UserId = int.Parse(userId)
         };
 
         var id = await _postService.CreatePostAsync(post);
@@ -63,11 +68,16 @@ public class PostsController : Controller
     [HttpPut(ApiRoutes.Posts.Update)]
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdatePostRequest request)
     {
-        var post = new Post
+        var userId = HttpContext.User.Claims.SingleOrDefault(u => u.Type == "id")?.Value;
+        var userOwnsPost = await _postService.UserOwnsPostAsync(id, int.Parse(userId));
+
+        if (!userOwnsPost)
         {
-            Id = id,
-            Name = request.Name
-        };
+            return BadRequest(new { error = "You do not own this post" });
+        }
+
+        var post = await _postService.GetPostByIdAsync(id);
+        post.Name = request.Name;
 
         var updated = await _postService.UpdatePostAsync(post);
 
@@ -82,6 +92,14 @@ public class PostsController : Controller
     [HttpDelete(ApiRoutes.Posts.Delete)]
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        var userId = HttpContext.User.Claims.SingleOrDefault(u => u.Type == "id")?.Value;
+        var userOwnsPost = await _postService.UserOwnsPostAsync(id, int.Parse(userId));
+
+        if (!userOwnsPost)
+        {
+            return BadRequest(new { error = "You do not own this post" });
+        }
+        
         var deleted = await _postService.DeletePostAsync(id);
 
         if (!deleted)
